+++ src/Slices/ServiceInstanceDetails/UI/Components/AttributesComponents/index.ts
@@ -1,2 +1,1 @@
 export * from "./AttributesViewProvider";
-export * from "./Utils";
