+++ src/Slices/ServiceInstanceDetails/UI/Components/InstanceActions/Actions/ExpertStateTransfer.tsx
@@ -13,7 +13,8 @@ import { uniqueId } from "lodash";
 import { ParsedNumber } from "@/Core";
 import { usePostExpertStateTransfer } from "@/Data/Managers/V2/POST/PostExpertStateTransfer";
 import { DependencyContext, words } from "@/UI";
-import { ConfirmationModal, ToastAlertMessage } from "../../Util";
+import { ConfirmationModal } from "../../ConfirmModal";
+import { ToastAlertMessage } from "../../ToastAllert";
 
 interface Props {
   targets: string[];
