--- src/Slices/ServiceInstanceDetails/UI/Components/Util/index.ts
@@ -1,2 +1,0 @@
-export * from "./ConfirmModal";
-export * from "./ToastAllert";
