+++ src/Slices/ServiceInstanceDetails/Utils/Utils.ts
